requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
